pycore/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vertile-api"
version = "0.1.0"
description = "Vertile FastAPI Backend"
requires-python = ">=3.8"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "websockets==12.0",
    "pydantic==2.4.2",
    "python-dotenv==1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "black>=22.1.0",
    "isort>=5.10.0",
    "mypy>=0.931",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 